Bugfix: Loopingcall was not terminated correctly.
Only the reactor was stopped but not the looping call. This apparently caused problems at the end of the program. The eventloop was rebuilt, so that now the loops are counted by a wrapper class and after n loops both the loopingcall and the reactor are stopped. ( Credits to: https://stackoverflow.com/questions/44968182/how-to-set-number-of-calls-of-the-function-scheduled-by-loopingcall)