Importing the temporal course of the measurement from an Excel file
The event loop for performing the interval measurements using twister and a reactor has been removed. Since the reactor can only be started once, it is not possible to iterate over the individual rows of the Excel table using the Wrapper class. The possibility to change the voltages during the measurement using an Excel table was requested by the student. Therefore, the function interval_measurement() was rebuilt to a while loop.